1.core模块下新增RestfulController，提取公共方法 2.简化common-captcha模块 3.集成redisson替代RedisTemplate，但是也内置了一个支持Jackson的RedisTemplate 4.引入sms4j 5.新增限流注解 6.图形验证码生成支持限流注解

+++ youlan-common/youlan-common-captcha/src/main/java/com/youlan/common/captcha/entity/ImageCaptcha.java
@@ -19,7 +19,7 @@ public class ImageCaptcha {
     private String captchaImg;
 
     @Schema(description = "验证码超时时间")
-    private Long codeTimeout;
+    private int codeTimeout;
 
     @Schema(description = "是否开启验证码")
     private Boolean captchaEnabled;
